Fix nav files sometimes not being generated by timer in time
Adding bots results in the game generating a nav file itself (which restarts the game if nav_restart_after_analysis is enabled and is generally annoying). Previously, bots would be added before timer-misc could generate the nav file, so the nav file generation was moved to just before bots are added instead.